fix(api-server,timeline-engine): replay 'main' label, type casts, COM bounds
- handlers/timeline::get_history: resolve well-known label "main" to
  BranchId::main() instead of generating a random UUID on parse failure.
  The frontend timeline panel sends "main" for the trunk; the previous
  fallback silently produced a UUID that matched no branch, so /api/
  timeline/{branch}/history always returned an empty list.
- timeline-engine::replay::ReplayError::KernelError: rename `source` to
  `message` (string, not Error). The thiserror `#[from]`-style `source`
  field name implied a real error trait object; this is a stringified
  message, and the field name was misleading callers into expecting
  Error::source() chaining.
- replay::dispatch_generic cylinder branch: cast `segments` to u32 (not
  usize) to match the kernel signature.
- protocol/geometry_handlers::estimate_center_of_mass_from_mesh: cast
  bounds midpoints to f64 explicitly. `Mesh::bounds()` now returns f32
  components after the f32-mesh tessellation work; the implicit f32→f64
  promotion silently dropped precision in the COM fallback path.

+++ roshera-backend/api-server/src/protocol/geometry_handlers.rs
@@ -1484,9 +1484,9 @@ fn estimate_center_of_mass_from_mesh(mesh: &Mesh) -> [f64; 3] {
     if total_volume.abs() < 1.0e-12 {
         let bounds = mesh.bounds();
         return [
-            (bounds.min[0] + bounds.max[0]) * 0.5,
-            (bounds.min[1] + bounds.max[1]) * 0.5,
-            (bounds.min[2] + bounds.max[2]) * 0.5,
+            ((bounds.min[0] + bounds.max[0]) * 0.5) as f64,
+            ((bounds.min[1] + bounds.max[1]) * 0.5) as f64,
+            ((bounds.min[2] + bounds.max[2]) * 0.5) as f64,
         ];
     }
